Remove RSpec3 deprecation warnings
